// frontend/app/[lang]/types/generated-strapi.ts
export interface StrapiError {
  status: number;
  name: string;
  message: string;
  details: Record<string, unknown>;
}

export interface StrapiPagination {
  start: number;
  limit: number;
  total: number;
}

export interface StrapiCollectionMeta {
  pagination: StrapiPagination;
}

export interface StrapiSingleTypeMeta {
  availableLocales?: string[];
}

export interface StrapiMedia {
  id: number;
  documentId: string;
  name: string;
  alternativeText?: string | null;
  caption?: string | null;
  width: number;
  height: number;
  formats?: {
    [key: string]: {
      name: string;
      hash: string;
      ext: string;
      mime: string;
      path: string | null;
      width: number;
      height: number;
      size: number;
      sizeInBytes: number;
      url: string;
    };
  } | null;
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  previewUrl?: string | null;
  provider: string;
  provider_metadata?: any | null;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
}

export type StrapiRelation<T> = {
  id: number;
  documentId: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
} & T;

export type StrapiCollection<T> = Array<{
  id: number;
  documentId: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
} & T>;

export interface StrapiResponse<T> {
  data: T | null;
  error?: StrapiError;
}
// ...
export interface Project {
  id: number;
  documentId: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  title: string;
  description: any;
  fullDescription: any;
  slug: string;
  linkToApp: string;
  linkToRepo: string;
  cover?: StrapiMedia;
  icon?: StrapiMedia;
}
// ...
export type ProjectSingleResponse = StrapiResponse<Project>;